requirements.txt
streamlit
matplotlib
plotly
pandas
psutil
google-generativeai
